+++ src/main/java/com/kobe/warehouse/service/reglement/dto/InvoicePaymentParam.java
@@ -2,4 +2,4 @@ package com.kobe.warehouse.service.reglement.dto;
 
 import java.time.LocalDate;
 
-public record InvoicePaymentParam(String search, Long organismeId, LocalDate dateDebut, LocalDate dateFin, boolean grouped) {}
+public record InvoicePaymentParam(String search, Integer organismeId, LocalDate dateDebut, LocalDate dateFin, boolean grouped) {}
